converted code injections to plain TS
